requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
cryptography==41.0.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
passlib[bcrypt]==1.7.4